Created abstract TokenAuthenticator and specific XSUAA/IAS authenticators.
